feat: implement StatCard component for reusable statistics display and update About page title refactor: enhance Contact form styles for better readability and user experience
