+++ agrilink/src/app/our-team/page.tsx
@@ -122,7 +122,7 @@ export default function About() {
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
           <div className="text-center">
             <h1 className="text-4xl md:text-6xl font-bold text-gray-900 mb-6">
-              About <span className="text-green-600">AgriLink</span>
+              About <span className="text-green-600">Us</span>
             </h1>
             <p className="text-xl text-gray-600 max-w-3xl mx-auto">
               Empowering Sri Lankan agriculture through intelligent technology and real-time market connections.
@@ -346,28 +346,7 @@ export default function About() {
             </div>
           </div>
 
-          {/* Our Impact Section */}
-          <div className="text-center">
-            <h2 className="text-3xl font-bold text-gray-900 mb-12">Our Impact</h2>
-            <div className="grid grid-cols-2 md:grid-cols-4 gap-8">
-              <div className="text-center">
-                <div className="text-3xl font-bold text-green-600 mb-2">50+</div>
-                <p className="text-gray-600">Districts Covered</p>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-blue-600 mb-2">100,000+</div>
-                <p className="text-gray-600">Price Updates Daily</p>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-green-600 mb-2">5,000+</div>
-                <p className="text-gray-600">Active Sellers</p>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-blue-600 mb-2">25,000+</div>
-                <p className="text-gray-600">Daily Users</p>
-              </div>
-            </div>
-          </div>
+          
         </div>
       </section>
